Do EDE encryption as in previous versions. Reseed the RNG when the program starts. Create both reserved records when building the database and not again later. Reserved records are marked private, which makes several things simpler. Boundary checks when unpacking in case we screwed up decryption. Properly store a SALT+VALUE hash. Keep the session key in the database and in the snib -- at the moment it is not encrypted. 16-byte lengths for record fields. Always show function name in system errors.

+++ keyring/record.h
@@ -24,7 +24,7 @@
 typedef struct {
     /* Length of corresponding string fields, not including the
      * terminating NUL that is present inside the memory blocks. */
-    UInt32 nameLen, acctLen, passwdLen, notesLen;
+    UInt16 nameLen, acctLen, passwdLen, notesLen;
     
     /* Handles to string values, or 0.
      *
